Refactor logical operators and remove conditional support
- Replaced the conditional operator with a logical 'if' operator, enhancing the logical evaluation framework.
- Updated the evaluator to handle the new 'if' operator, ensuring proper argument validation and evaluation.
- Removed the conditional operators module and related tests, streamlining the logic module.
- Adjusted tests to reflect changes in operator types and ensure compatibility with the new logical structure.

These changes aim to simplify the operator model and improve the clarity and functionality of logical operations.
