+++ src/logic/operators/mod.rs
@@ -8,7 +8,6 @@ pub mod logical;
 pub mod string;
 pub mod missing;
 pub mod array;
-pub mod conditional;
 pub mod log;
 pub mod r#in;
 pub mod variable;
@@ -19,4 +18,3 @@ pub use arithmetic::ArithmeticOp;
 pub use logical::LogicalOp;
 pub use string::StringOp;
 pub use array::ArrayOp;
-pub use conditional::ConditionalOp;
